Add files via upload
This initial commit introduces the core simulation notebooks for the Master's thesis project on modeling neuronal excitability. The first notebook implements a Hodgkin-Huxley model to analyze excitability regimes in the S-K parameter space, while the second provides a comparative biophysical model of young versus mature neurons based on existing literature. This repository is a work in progress and will be expanded with a detailed README file, figures, and additional analysis in subsequent commits.


References :
https://pubmed.ncbi.nlm.nih.gov/1607940/
https://www.pnas.org/doi/10.1073/pnas.1808552115